+++ nextgisweb/pyramid/nodepkg/hook/useRouteGet.js
@@ -1,9 +1,19 @@
 import { useEffect, useState } from "react";
 import { useRoute } from "./useRoute";
 
-export function useRouteGet({ name, params = {}, options = {} }) {
-    params =  { loadOnInit: true, ...params };
-    const { route, abort } = useRoute(name, params);
+export function useRouteGet(nameOrProps, params = {}, options = {}) {
+    let endpointName = "";
+
+    if (typeof nameOrProps === "string") {
+        endpointName = nameOrProps;
+    } else {
+        endpointName = nameOrProps.name;
+        params = { ...nameOrProps.params, ...params };
+        options = { ...nameOrProps.options, ...options };
+    }
+
+    params = { loadOnInit: true, ...params };
+    const { route, abort } = useRoute(endpointName, params);
     const [isLoading, setIsLoading] = useState(!!params.loadOnInit);
     const [data, setData] = useState();
     const [error, setError] = useState();