+++ nextgisweb/webmap/nodepkg/clone-webmap/index.js
@@ -1,0 +1,3 @@
+/** @entrypoint */
+import { CloneWebmap } from "./CloneWebmap";
+export default CloneWebmap;