+++ CHANGES.rst
@@ -6,6 +6,7 @@ Changes
 
 - OpenLayers library upgraded to 6.15.1.
 - Use resource SRS by default while exporting raster and vector layers.
+- Support for webmap cloning through UI.
 
 
 4.2.0